# Improve BlockDist scan scalability