--- a/test/modules/bradc/printModStuff/foo.good
+++ b/test/modules/bradc/printModStuff/foo.good
@@ -48,7 +48,6 @@ end of module search dirs
   $CHPL_HOME/modules/standard/Time.chpl
   $CHPL_HOME/modules/layouts/LayoutCS.chpl
   $CHPL_HOME/modules/dists/SparseBlockDist.chpl
-  $CHPL_HOME/modules/dists/ReplicatedDist.chpl
   $CHPL_HOME/modules/standard/Memory/Diagnostics.chpl
   $CHPL_HOME/modules/standard/Memory/Initialization.chpl
   $CHPL_HOME/modules/standard/Memory.chpl