@bradcray 2022-06-13
I put some comments and thoughts, though none of them are show-stoppers.  Probably the one that I'd do if it were my code would be to change the var, nilable, unmanaged class into a const, non-nilable, owned class, taking a borrow to it for the array to store.